Move usePrometheusPoll hook to dynamic plugin SDK
