+++ frontend/packages/console-dynamic-plugin-sdk/@types/sdk/index.d.ts
@@ -1,6 +1,8 @@
 declare interface Window {
   SERVER_FLAGS: {
     basePath: string;
+    prometheusBaseURL: string;
+    prometheusTenancyBaseURL: string;
   };
   windowError?: string;
   __REDUX_DEVTOOLS_EXTENSION_COMPOSE__?: Function;
