+++ frontend/packages/console-dynamic-plugin-sdk/src/lib-core.ts
@@ -1,2 +1,3 @@
 export * from './extensions';
 export * from './api/core-api';
+export * from './api/prometheus-types';
