+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/fetch/hooks/useSafeFetch.ts
@@ -1,0 +1,13 @@
+import { useEffect, useRef } from 'react';
+import { consoleFetchJSON } from '../console-fetch';
+
+export const useSafeFetch = () => {
+  const controller = useRef<AbortController>();
+  useEffect(() => {
+    controller.current = new AbortController();
+    return () => controller.current.abort();
+  }, []);
+
+  return (url) =>
+    consoleFetchJSON(url, 'get', { signal: controller.current.signal as AbortSignal });
+};
