+++ frontend/packages/console-dynamic-plugin-sdk/src/utils/prometheus/utils.ts
@@ -1,0 +1,58 @@
+import { PrometheusEndpoint } from '../../api/prometheus-types';
+import { ONE_HOUR } from '../time';
+
+export const PROMETHEUS_BASE_PATH = window.SERVER_FLAGS.prometheusBaseURL;
+export const PROMETHEUS_TENANCY_BASE_PATH = window.SERVER_FLAGS.prometheusTenancyBaseURL;
+export const DEFAULT_PROMETHEUS_SAMPLES = 60;
+export const DEFAULT_PROMETHEUS_TIMESPAN = ONE_HOUR;
+
+// Range vector queries require end, start, and step search params
+const getRangeVectorSearchParams = (
+  endTime: number = Date.now(),
+  samples: number = DEFAULT_PROMETHEUS_SAMPLES,
+  timespan: number = DEFAULT_PROMETHEUS_TIMESPAN,
+): URLSearchParams => {
+  const params = new URLSearchParams();
+  params.append('start', `${(endTime - timespan) / 1000}`);
+  params.append('end', `${endTime / 1000}`);
+  params.append('step', `${timespan / samples / 1000}`);
+  return params;
+};
+
+const getSearchParams = ({
+  endpoint,
+  endTime,
+  timespan,
+  samples,
+  ...params
+}: PrometheusURLProps): URLSearchParams => {
+  const searchParams =
+    endpoint === PrometheusEndpoint.QUERY_RANGE
+      ? getRangeVectorSearchParams(endTime, samples, timespan)
+      : new URLSearchParams();
+  Object.keys(params).forEach(
+    (param) => params[param] && searchParams.append(param, params[param].toString()),
+  );
+  return searchParams;
+};
+
+export const getPrometheusURL = (
+  props: PrometheusURLProps,
+  basePath: string = props.namespace ? PROMETHEUS_TENANCY_BASE_PATH : PROMETHEUS_BASE_PATH,
+): string => {
+  if (props.endpoint !== PrometheusEndpoint.RULES && !props.query) {
+    return '';
+  }
+  const params = getSearchParams(props);
+  return `${basePath}/${props.endpoint}?${params.toString()}`;
+};
+
+type PrometheusURLProps = {
+  endpoint: PrometheusEndpoint;
+  endTime?: number;
+  namespace?: string;
+  query?: string;
+  samples?: number;
+  timeout?: string;
+  timespan?: number;
+};
