+++ frontend/packages/console-shared/src/constants/index.ts
@@ -2,5 +2,4 @@ export * from './pod';
 export * from './resource';
 export * from './common';
 export * from './ui';
-export * from './time';
 export * from './duration';
