+++ frontend/packages/console-shared/src/constants/time.ts
@@ -1,5 +1,1 @@
-export const ONE_SECOND = 1000;
-export const ONE_MINUTE = 60 * ONE_SECOND;
-export const ONE_HOUR = 60 * ONE_MINUTE;
-export const ONE_DAY = 24 * ONE_HOUR;
-export const ONE_WEEK = 7 * ONE_DAY;
+export * from '@console/dynamic-plugin-sdk/src/utils/time';
