+++ frontend/public/components/build.tsx
@@ -6,7 +6,8 @@ import * as classNames from 'classnames';
 import { sortable } from '@patternfly/react-table';
 import { Alert, Grid, GridItem } from '@patternfly/react-core';
 
-import { ONE_HOUR, ONE_MINUTE, Status } from '@console/shared';
+import { ONE_HOUR, ONE_MINUTE } from '@console/dynamic-plugin-sdk/src/utils/time';
+import { Status } from '@console/shared';
 import { ByteDataTypes } from '@console/shared/src/graph-helper/data-utils';
 import {
   K8sResourceKindReference,
